requirements.txt
tornado
requests
simplejson
rjsmin
rcssmin
